add async method for URLSession, swiftlint, logger and FileCache to package
