+++ .swiftlint.yml
@@ -1,0 +1,30 @@
+excluded: # case-sensitive paths to ignore during linting. Takes precedence over `included`
+  - Pods
+disabled_rules:
+  - block_based_kvo
+  - computed_accessors_order
+  - cyclomatic_complexity
+  - file_length
+  - force_cast
+  - force_try
+  - function_body_length
+  - function_parameter_count
+  - identifier_name
+  - implicit_getter
+  - inclusive_language
+  - large_tuple
+  - line_length
+  - multiple_closures_with_trailing_closure
+  - orphaned_doc_comment
+  - private_over_fileprivate
+  - todo
+  - trailing_comma
+  - trailing_whitespace
+  - type_body_length
+  - type_name
+  - xctfail_message
+  - unavailable_condition
+  - void_function_in_ternary
+  - redundant_objc_attribute
+  - blanket_disable_command
+
