+++ Package.swift
@@ -1,0 +1,23 @@
+// swift-tools-version: 5.10
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "ToDoShka",
+    products: [
+        // Products define the executables and libraries a package produces, making them visible to other packages.
+        .library(
+            name: "ToDoShka",
+            targets: ["ToDoShka"]),
+    ],
+    targets: [
+        // Targets are the basic building blocks of a package, defining a module or a test suite.
+        // Targets can depend on other targets in this package and products from dependencies.
+        .target(
+            name: "ToDoShka"),
+        .testTarget(
+            name: "ToDoShkaTests",
+            dependencies: ["ToDoShka"]),
+    ]
+)
