+++ Podfile
@@ -1,0 +1,16 @@
+# Uncomment the next line to define a global platform for your project
+# platform :ios, '17.0'
+
+target 'ToDoShka' do
+  # Comment the next line if you don't want to use dynamic frameworks
+  use_frameworks!
+
+  # Pods for ToDoShka
+  pod 'SwiftLint'
+
+  target 'ToDoShkaTests' do
+    inherit! :search_paths
+    # Pods for testing
+  end
+
+end
