+++ ToDoShka/ToDoShkaApp.swift
@@ -1,11 +1,15 @@
 import SwiftUI
+import CocoaLumberjackSwift
 
 @main
 struct ToDoShkaApp: App {
     
     var body: some Scene {
         return WindowGroup {
-            ContentView()
+
+            ContentView().onAppear {
+                DDLog.add(DDTTYLogger.sharedInstance!)
+            }
         }
     }
 }
